requirements.txt

psutil
pywin32
Pillow
pynput
numpy
